requirements.txt
streamlit==1.21.0
pinecone-client==2.1.0
PyMuPDF==1.21.0
python-dotenv==1.0.0
groq==0.1.0
numpy==1.24.3
pandas==2.0.2
Pillow==9.4.0
requests==2.31.0
